Changed MPDSDataRetrieval setting to be per instance
In the methods the following settings:
    endpoint = "https://api.mpds.io/v0/download/facet"

    pagesize = 1000
    maxnpages = 100   # one hit may reach 50kB in RAM, consider pagesize*maxnpages*50kB free RAM
    maxnphases = 1500 # more phases require additional requests
    chillouttime = 3  # please, do not use values < 3, because the server may burn out

were being accessed using the class, i.e., MPDSDataRetrieval.pagesize.

This means, if the user wants to change it they have to do it on the
class.  This is undesirable and should be on a per instance basis.
The user may legimately have two (or more) MPDSDataRetrieval instances
and these could have different settings e.g.:

r1 = MPDSDataRetrieval(api_key=...)
r1.pagesize = 10
r2 = MPDSDataRetrieval(api_key=...)
r2.pagesize = 100

Before this commit neither of the assignments of pagesize would have an
affect and the class level pagesize of 1000 would remain in place.  To
me this is unexpected behaviour and this commit 'fixes' it.